consumer/requirements.txt
kafka-python>=2.0.2,<3.0
